Convert described form group to typeScript
